requirements.txt
pyodbc
flask
azure-storage-blob
azure-identity